% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.importexport_operations.R
\name{cancel_job}
\alias{cancel_job}
\title{This operation cancels a specified job}
\usage{
cancel_job(JobId, APIVersion = NULL)
}
\arguments{
\item{JobId}{[required]}

\item{APIVersion}{}
}
\description{
This operation cancels a specified job. Only the job owner can cancel it. The operation fails if the job has already started or is complete.
}
\section{Accepted Parameters}{
\preformatted{cancel_job(
  JobId = "string",
  APIVersion = "string"
)
}
}

