% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.importexport_operations.R
\name{get_shipping_label}
\alias{get_shipping_label}
\title{This operation generates a pre-paid UPS shipping label that you will use to ship your device to AWS for processing}
\usage{
get_shipping_label(jobIds, name = NULL, company = NULL,
  phoneNumber = NULL, country = NULL, stateOrProvince = NULL,
  city = NULL, postalCode = NULL, street1 = NULL, street2 = NULL,
  street3 = NULL, APIVersion = NULL)
}
\arguments{
\item{jobIds}{}

\item{name}{}

\item{company}{}

\item{phoneNumber}{}

\item{country}{}

\item{stateOrProvince}{}

\item{city}{}

\item{postalCode}{}

\item{street1}{}

\item{street2}{}

\item{street3}{}

\item{APIVersion}{}
}
\description{
This operation generates a pre-paid UPS shipping label that you will use to ship your device to AWS for processing.
}
\examples{

}
