% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.importexport_package.R
\docType{package}
\name{paws.importexport-package}
\alias{paws.importexport}
\alias{paws.importexport-package}
\title{paws.importexport: AWS Import/Export}
\description{
AWS Import/Export Service AWS Import/Export accelerates
transferring large amounts of data between the AWS cloud and portable
storage devices that you mail to us. AWS Import/Export transfers data
directly onto and off of your storage devices using Amazon\'s
high-speed internal network and bypassing the Internet. For large data
sets, AWS Import/Export is often faster than Internet transfer and
more cost effective than upgrading your connectivity.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
