% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.importexport_operations.R
\name{update_job}
\alias{update_job}
\title{You use this operation to change the parameters specified in the original manifest file by supplying a new manifest file}
\usage{
update_job(JobId, Manifest, JobType, ValidateOnly, APIVersion = NULL)
}
\arguments{
\item{JobId}{}

\item{Manifest}{}

\item{JobType}{}

\item{ValidateOnly}{}

\item{APIVersion}{}
}
\description{
You use this operation to change the parameters specified in the original manifest file by supplying a new manifest file. The manifest file attached to this request replaces the original manifest file. You can only use the operation after a CreateJob request but before the data transfer starts and you can only use it on jobs you own.
}
\examples{

}
