% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.importexport_operations.R
\name{get_shipping_label}
\alias{get_shipping_label}
\title{This operation generates a pre-paid UPS shipping label that you will use to ship your device to AWS for processing}
\usage{
get_shipping_label(
  jobIds = list(
    "string"
  ),
  name = "string",
  company = "string",
  phoneNumber = "string",
  country = "string",
  stateOrProvince = "string",
  city = "string",
  postalCode = "string",
  street1 = "string",
  street2 = "string",
  street3 = "string",
  APIVersion = "string"
)
}
\arguments{
\item{jobIds}{[required]}

\item{name}{}

\item{company}{}

\item{phoneNumber}{}

\item{country}{}

\item{stateOrProvince}{}

\item{city}{}

\item{postalCode}{}

\item{street1}{}

\item{street2}{}

\item{street3}{}

\item{APIVersion}{}
}
\description{
This operation generates a pre-paid UPS shipping label that you will use to ship your device to AWS for processing.
}
\examples{

}
