% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.importexport_operations.R
\name{get_status}
\alias{get_status}
\title{This operation returns information about a job, including where the job is in the processing pipeline, the status of the results, and the signature value associated with the job}
\usage{
get_status(JobId, APIVersion = NULL)
}
\arguments{
\item{JobId}{}

\item{APIVersion}{}
}
\description{
This operation returns information about a job, including where the job is in the processing pipeline, the status of the results, and the signature value associated with the job. You can only return information about jobs you own.
}
\examples{

}
