% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.importexport_operations.R
\name{update_job}
\alias{update_job}
\title{You use this operation to change the parameters specified in the original manifest file by supplying a new manifest file}
\usage{
update_job(JobId, Manifest, JobType, ValidateOnly, APIVersion = NULL)
}
\arguments{
\item{JobId}{[required]}

\item{Manifest}{[required]}

\item{JobType}{[required]}

\item{ValidateOnly}{[required]}

\item{APIVersion}{}
}
\description{
You use this operation to change the parameters specified in the original manifest file by supplying a new manifest file. The manifest file attached to this request replaces the original manifest file. You can only use the operation after a CreateJob request but before the data transfer starts and you can only use it on jobs you own.
}
\section{Accepted Parameters}{
\preformatted{update_job(
  JobId = "string",
  Manifest = "string",
  JobType = "Import"|"Export",
  ValidateOnly = TRUE|FALSE,
  APIVersion = "string"
)
}
}

